pyproject.toml
[tool.poetry]
name = "spreadsheet-to-page-in-db"
version = "0.1.0"
description = ""
authors = ["KentaroKajiyama <kemushi555@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"
rich = "^13.9.4"
setuptools = "^75.8.0"
main = {path = "spreadsheet_to_page_in_db"}
markdown-it-py = "^3.0.0"
mdit-py-plugins = "^0.4.2"
linkify-it-py = "^2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
